changing all about configuration in defs.c/.h
