+++ STM32CubeIDE/src/Board/Interface/stm32g0xx_hal_msp.c
@@ -60,19 +60,6 @@
                     /**
   * Initializes the Global MSP.
   */
-void HAL_MspInit(void)
-{
-
-  /* USER CODE BEGIN MspInit 0 */
-
-  /* USER CODE END MspInit 0 */
-
-  /* System interrupt init*/
-
-  /* USER CODE BEGIN MspInit 1 */
-
-  /* USER CODE END MspInit 1 */
-}
 
 /**
   * @brief ADC MSP Initialization
